+++ bootstrap.ps1
@@ -30,16 +30,7 @@ param (
     [string] $BetaEcrs,
 
     [Parameter(Mandatory = $false, HelpMessage = "Semicolon separated ECR URIs to store Prod images.")]
-    [string] $ProdEcrs,
-
-    [Parameter(Mandatory = $true, HelpMessage = "ECR repository name for Stage, Beta and Prod images.")]
-    [string] $EcrRepositoryName,
-
-    [Parameter(Mandatory = $true, HelpMessage = "The target .NET framework to create a pipeline for.")]
-    [string] $TargetFramework,
-
-    [Parameter(Mandatory = $true, HelpMessage = "The .NET channel corresponding to the Target Framework.")]
-    [string] $DotnetChannel
+    [string] $ProdEcrs
 )
 
 $env:AWS_LAMBDA_PIPELINE_ACCOUNT_ID = $PipelineAccountId
@@ -56,11 +47,6 @@ $env:AWS_LAMBDA_GITHUB_REPO_BRANCH = $GitHubRepoBranch
 $env:AWS_LAMBDA_STAGE_ECR = $StageEcr
 $env:AWS_LAMBDA_BETA_ECRS = $BetaEcrs
 $env:AWS_LAMBDA_PROD_ECRS = $ProdEcrs
-
-$env:AWS_LAMBDA_ECR_REPOSITORY_NAME = $EcrRepositoryName
-
-$env:AWS_LAMBDA_DOTNET_FRAMEWORK_VERSION = $TargetFramework
-$env:AWS_LAMBDA_DOTNET_FRAMEWORK_CHANNEL = $DotnetChannel
 
 npx cdk bootstrap --cloudformation-execution-policies arn:aws:iam::aws:policy/AdministratorAccess aws://$PipelineAccountId/$Region
 npx cdk deploy --require-approval never --all